Version qui bugue pas

+++ LivrOnTime/src/main/java/model/Plan.java
@@ -186,7 +186,7 @@ public Tournee calculerLaTournee(DemandeLivraison dl) {
 	long[][] time = new long[dl.getLivraisons().size()+1][2];
 	for (int i= 0;i<dl.getLivraisons().size();i++) {
 		time[i][0]=(dl.getLivraisons().get(i).getDebutPlageHoraire()==null? -1:dl.getLivraisons().get(i).getDebutPlageHoraire().getTime()-dl.getHeureDepart().getTime());
-		time[i][1]=(dl.getLivraisons().get(i).getFinPlageHoraire()==null? Integer.MAX_VALUE:dl.getLivraisons().get(i).getFinPlageHoraire().getTime()-dl.getHeureDepart().getTime());
+		time[i][1]=(dl.getLivraisons().get(i).getFinPlageHoraire()==null? Long.MAX_VALUE:dl.getLivraisons().get(i).getFinPlageHoraire().getTime()-dl.getHeureDepart().getTime());
 	}
 	
 	etape2.chercheSolution(tpLimite,nbSommet, graphe_complet.getCout(),graphe_complet.getDuree(), time);
